ext_external_objects/vk-stencil-display: New interoperability test
The purpose of this test is to use a Vulkan stencil buffer from OpenGL.
The test renders a quad with Vulkan and sets the stencil value to 1
for every pixel that is rendered. The stencil buffer is then imported in
OpenGL and is used as the depth/stencil attachment of an OpenGL FBO. The
color attachment of the FBO is cleared to green and is used as render
target where a blue quad is rendered with the stencil test enabled and a
stencil mask/reference combination that permits drawing only where the
stencil value is 0. The final result must be a green quad with a smaller
blue quad in the middle.

v2:  fixed commit message
v3:  fix in probe (the colors were reversed accidentally)
v4:  special case for AMD that doesn't support 24/8 depth/stencil
v5:  removed special case and modified the test to run subtests per
format which are skipped if a format is not supported
v6:  removed empty space
v7:  renamed all_formats to depth_stencil_formats that is more
meaningful, removed trailing whitespaces
v8:  changed the loop of stencil subtests to skip the depth-only format
v9:  fixed indentation
v10: added base and max level (TexParameteri) for OpenGL color texture,
     fixes incomplete_tex mesa error (MESA_DEBUG=incomplete_tex)
v11: added the test to opengl.py
v12: removed double depth test state change, and disabled the stencil
right after rendering to its corresponding FBO
v13: added stencilLoadOp, stencilStoreOp values in create_renderpass in
vk.c
v14: removed state change from gl_subtest_init since we re-change it in
run_subtest right afterwards

Signed-off-by: Eleni Maria Stea <estea@igalia.com>
Reviewed-by: Tapani Pälli <tapani.palli@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/508>