glsl: Test out-of-bounds access to array of opaque types
From section 5.7 "Structure and Array Operations" of the GLSL 1.30 spec:

 "Behavior is undefined if a shader subscripts an array with an index less
  than 0 or greater than or equal to the size the array was declared with."

The behaviour becomes defined only in robustness extensions, however even
if driver is technically allowed to crash or hang, it most likely
doesn't want to.

Robustness extensions encourage to us to return zero on out-of-bounds
read, however it's not clearly applicable to the array of opaque types.
What will happen when we call a built-in function which expects the
opaque type with the instance returned from oob access?
In the spirit of the spec - I think neither compilation should fail nor
a hang occur.

Signed-off-by: Danylo Piliaiev <danylo.piliaiev@globallogic.com>
Acked-by: Emma Anholt <emma@anholt.net>
Reviewed-by: Ian Romanick <ian.d.romanick@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/373>