tests: Add test to stress Mesa Gallium draw module.
This test was designed to exercise as many of the Mesa Gallium draw
module's vertex pipeline stages as possible, particularly those which
allocate extra vertex attributes.

It does not yet check for correctness, and is instead meant to be used
with memory debuggers to trap buffer overflows.

Reviewed-by: Roland Scheidegger <sroland@vmware.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/531>