glx_arb_create_context: fix no error test and check error value
The first test was always skipped because we were looking for an
error value of 0 for the positive test be the value is initialised
to -1 not 0.

We add a new check to make sure the error value is what the spec
says it should be. Note this also resets the error value which
was not being done before either.

Mesa currently fails this test as the xserver doesn't validate the
errors at all.

Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/548>