glsl-1.20: Test various aspects of do-while loop scoping
This test also inadvertently tests the sequence operator.  At the time
this test was written, Mesa would not compile it.  See also piglit!569.
glslangValidator also fails to compile it (see
https://github.com/KhronosGroup/glslang/issues/2744), but NVIDIA's
closed-source driver does compile it.

Once the compilation problem is resolved, I would like to wait to land
this test until we are at least sure it won't lead to a GPU hang.  The
original version of this test would, but this version should be safe.
Be making various tweaks to the test (adding { } around the body of the
loop, s/int i/i/ inside the loop), I'm pretty confident that it tests
everything the way I intend without causing an infinite loop.

Reviewed-by: Timothy Arceri <tarceri@itsqueeze.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/575>