arb_shader_image_load_store/*: Fix unintended ARB_viewport_array requirement.
We only have a single viewport, so no need to use the new interface
for specifying it.  Fixes a flake on freedreno where a compiler
assertion fail in the compiler thread raced against the skip when we
accidentally stumbled across the missing extension.

Reviewed-by: Ilia Mirkin <imirkin@alum.mit.edu>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/597>