framework/replay: Add argument to receive a JWT token as a file
Currently, piglit replay has the -j/--jwt argument to receive the token
directly to access minio server.
There is a use case where the JWT should not be exposed, where the token
per se should not be shown in the command line. So giving a file path
instead of the token is a better approach to avoid this leakage.

This commit adds --jwt-file argument as an alternative to -j/--jwt one.
When the user gives this argument, piglit will open it and read the file
content, treating it as the JWT token.

Reviewed-by: Tomeu Vizoso <tomeu.vizoso@collabora.com>
Reviewed-by: Cristian Ciocaltea <cristian.ciocaltea@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/605>