Tests for variable in and out of loop scope
Tests are helpful for understanding behaviour of variables
with different initialization and declaration places.
This may prevent from semantic errors in future.
Test are helpful for work with compiler components, tokens, etc.
Because it will check if the right behaviour of variables
was not suddenly changed.
More detailed information can be observed here:
https://gitlab.freedesktop.org/mesa/mesa/-/issues/5367

Signed-off-by: Viktoriia Palianytsia <v.palianytsia@globallogic.com>
Reviewed-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/616>