ext_external_objects: modify vk_vert_buf_update_errors test
Earlier expectation of this test was that glBufferSubData is not
supposed to work since mapping the buffer is prohibited by the spec.
However while spec explicitly prohibits mapping via MapBufferRange
and MapNamedBufferRange, it does not prohibit buffer updates via
glBufferSubData.

This commit changes test to create empty buffer at VK side, import
that to GL, update it with checkerboard pattern, render with buffer
to the display and verify results.

Signed-off-by: Tapani Pälli <tapani.palli@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/636>