ext_external_objects: modify vk_pix_buf_update_errors test
Earlier expectation of this test was that glBufferSubData is not
supposed to work since mapping the buffer is prohibited by the spec.
However while spec explicitly prohibits mapping via MapBufferRange
and MapNamedBufferRange, it does not prohibit buffer updates via
glBufferSubData.

This commit changes test to render reversed bands pattern with
Vulkan. In GL side, we update buffer with glBufferSubData to the
expected pattern, render with buffer to the display and verify
results match expectation.

Signed-off-by: Tapani Pälli <tapani.palli@intel.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/636>