glx-multithread-texture: Avoid front-buffer rendering.
It's not what this test is about, and it means that you need a
compositor for reliable results when run in parallel with other tests.

While I'm here, drop the extra swapbuffers for each frame of correct
rendering, and just swap the last one.  This should encourage losing
multithread races while making the test more efficient.

Reviewed-by: Marek Olšák <marek.olsak@amd.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/639>