framework/replay: display whole image after clicking on it
When we use right button and "Open Image in New Tab", sadly minio offers
the image only to download - not to look at. This non-linearity when
comparing the images for me.

This behaviour shows full-size image directly in browser when the image is clicked on.

Signed-off-by: David Heidelberg <david.heidelberg@collabora.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/697>