framework/replay: Add support for the new S3 servers
The new s3 server (https://s3.freedesktop.org) is using a different
authorization mechanism. We don't need to pre-login anymore with it,
we just need to forward the JWT token in every request.

Add support of it so we can retire the old MinIO server.

Signed-off-by: Benjamin Tissoires <benjamin.tissoires@gmail.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/744>