tests: add test that validates smooth-line coverage
This tests drawing a horizontal smooth line divided into varying length
segments, using additive blending to validate that the sum of the
coverage adds up to one, as one would expect.

While this isn't an explicit requirement in the OpenGL spec, the
identity should be a requirement due to the way smooth lines is
specified. It's also something that most reasonable approximations
should be able to produce.

Reviewed-by: Roland Scheidegger <sroland@vmware.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/747>