tests: add a test for the combination of smooth+stipple
So far, there's no tests testing the combination of GL_LINE_SMOOTH and
GL_LINE_STIPPLE. Let's add one, that draws a very basic stipple-pattern
twice, once normal and then once bitwise inverted. For each line we
draw, we shift the pattern by a quarter of a pixel.

This will show that the stipple pattern is applied "before" the
line-smoothing, i.e that the line-segments that are produced by the
stippling is individually smooth.

Reviewed-by: Roland Scheidegger <sroland@vmware.com>
Part-of: <https://gitlab.freedesktop.org/mesa/piglit/-/merge_requests/747>